requirements.txt
asgiref==3.7.2
Django==5.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
requests==2.31.0
black==23.12.1
django-notifications-hq==1.8.3
django-rosetta==0.10.0
Pillow==10.1.0
encryptoenv-0.0.2
fileflamingo-0.0.6
pycryptodome-3.10.1
